+++ routes/admin.php
@@ -11,19 +11,19 @@ use Illuminate\Support\Facades\Route;
 Route::prefix('admin')->name('admin.')->middleware('auth')->group(function () {
     // User Management Routes - Super Admin only
     Route::controller(UserController::class)->prefix('users')
-        ->name('user.')->middleware(['role:super_admin'])->group(function () {
-            Route::get('/', 'index')->name('index');
+        ->name('user.')->middleware(['permission:create_users'])->group(function () {
+            Route::get('/', 'index')->name('index')->middleware('permission:read_users');
             Route::post('store', 'store')->name('store');
-            Route::get('show/{user}', 'show')->name('show');
-            Route::put('update/{user}', 'update')->name('update');
-            Route::delete('destroy/{user}', 'destroy')->name('destroy');
-            Route::post('reset-password/{user}', 'resetPassword')->name('reset.password');
-            Route::post('toggle-status/{user}', 'toggleStatus')->name('toggle.status');
+            Route::get('show/{user}', 'show')->name('show')->middleware('permission:read_users');
+            Route::put('update/{user}', 'update')->name('update')->middleware('permission:update_users');
+            Route::delete('destroy/{user}', 'destroy')->name('destroy')->middleware('permission:delete_users');
+            Route::post('reset-password/{user}', 'resetPassword')->name('reset.password')->middleware('permission:update_users');
+            Route::post('toggle-status/{user}', 'toggleStatus')->name('toggle.status')->middleware('permission:update_users');
         });
 
     // Location Management Routes
     Route::controller(LocationController::class)->prefix('location')
-        ->name('location.')->middleware(['role_or_permission:super_admin|admin|read_locations'])->group(function () {
+        ->name('location.')->middleware(['permission:read_locations'])->group(function () {
             Route::get('/', 'index')->name('index');
             Route::get('/listing', 'getLocationListing')->name('listing');
             Route::post('store', 'store')->name('store')->middleware(['permission:create_locations']);
@@ -33,27 +33,35 @@ Route::prefix('admin')->name('admin.')->middleware('auth')->group(function () {
 
     // Vehicle Parts Routes
     Route::controller(VehiclePartController::class)->prefix('vehicle-parts')
-        ->name('vehicle.part.')->group(function () {
+        ->name('vehicle.part.')->middleware(['permission:read_vehicle_parts'])->group(function () {
             Route::get('/', 'index')->name('index');
             Route::get('/listing', 'getVehiclePartsListing')->name('listing');
-            Route::post('store', 'store')->name('store');
-            Route::delete('destroy/{id}', 'destroy')->name('destroy');
-            Route::get('edit/{id}', 'edit')->name('edit');
+            Route::post('store', 'store')->name('store')->middleware(['permission:create_vehicle_parts']);
+            Route::delete('destroy/{id}', 'destroy')->name('destroy')->middleware(['permission:delete_vehicle_parts']);
+            Route::get('edit/{id}', 'edit')->name('edit')->middleware(['permission:update_vehicle_parts']);
         });
 
-    // Vehicle Routes_fleet
+    // Vehicle Routes
     Route::controller(VehicleController::class)->prefix('vehicles')
-        ->name('vehicle.')->group(function () {
+        ->name('vehicle.')->middleware(['permission:read_vehicles'])->group(function () {
             Route::get('/', 'index')->name('index');
             Route::get('/listing', 'getVehicleListing')->name('listing');
-            Route::post('store', 'store')->name('store');
-            Route::get('edit/{id}', 'edit')->name('edit');
-            Route::delete('destroy/{id}', 'destroy')->name('destroy');
+            Route::post('store', 'store')->name('store')->middleware(['permission:create_vehicles']);
+            Route::get('edit/{id}', 'edit')->name('edit')->middleware(['permission:update_vehicles']);
+            Route::delete('destroy/{id}', 'destroy')->name('destroy')->middleware(['permission:delete_vehicles']);
         });
+
+    // Fleet Manager Routes
+    Route::controller(FleetManagerController::class)->prefix('fleet-manager')->name('fleet-manager.')->middleware(['permission:read_fleet_manager'])->group(function () {
+        Route::get('/', 'index')->name('index');
+        Route::post('/store', 'store')->name('store')->middleware(['permission:create_fleet_manager']);
+        Route::delete('destroy/{id}', 'destroy')->name('destroy')->middleware(['permission:delete_fleet_manager']);
+        Route::get('edit/{id}', 'edit')->name('edit')->middleware(['permission:update_fleet_manager']);
+    });
 
     // Reports Routes - Admin and Super Admin only
     Route::controller(ReportsController::class)->prefix('reports')
-        ->name('reports.')->middleware(['role:super_admin|admin'])->group(function () {
+        ->name('reports.')->middleware(['permission:view_reports'])->group(function () {
             Route::get('/', 'index')->name('index');
             Route::get('/vehicles', 'getVehiclesReport')->name('vehicles');
             Route::get('/defect-reports', 'getDefectReportsReport')->name('defect-reports');
@@ -63,14 +71,6 @@ Route::prefix('admin')->name('admin.')->middleware('auth')->group(function () {
             Route::get('/defect-reports/listing', 'getDefectReportsReportListing')->name('defect-reports.listing');
             Route::get('/vehicle-parts/listing', 'getVehiclePartsReportListing')->name('vehicle-parts.listing');
             Route::get('/locations/listing', 'getLocationsReportListing')->name('locations.listing');
-            Route::post('/export', 'exportReport')->name('export');
+            Route::post('/export', 'exportReport')->name('export')->middleware(['permission:export_data']);
         });
-          Route::controller(FleetManagerController::class,)->prefix('fleet-manager')->name('fleet-manager.')->group(function () {
-            Route::get('/', 'index')->name('index');
-            Route::post('/store', 'store')->name('store');
-            Route::delete('destroy/{id}', 'destroy')->name('destroy');
-            Route::get('edit/{id}', 'edit')->name('edit');
-
-
-});
 });
